RK {rkward}	R Documentation

Plot graphics to RKWard native device

Description

RK() creates an R on-screen device that will be rendered in the RKWard frontend. The default settings for width, and height can be modified from Settings -> Configure RKWard -> Onscreen Graphics.

Usage

RK(width = getOption("rk.screendevice.width"),
  height = getOption("rk.screendevice.height"), pointsize = 12,
  family = NULL, bg = "white", title = "", antialias = TRUE)

Arguments

width
Width of the device in inches. The default is to use the heigth configured in Settings -> Configure RKWard -> Onscreen Graphics.

height
Height of the device in inchesgraphics in pixels. The default is to use the heigth configured in Settings -> Configure RKWard -> Onscreen Graphics.

pointsize
Default pointsize

family
Default font family. This can be a character vector of length 1 or 2. The second value is used for plotting symbols. Effectively the default is c("Helvetica", "Symbol"). A wide variety of sepcification is supported, including the device independent fonts names "sans", "serif", and "mono"

bg
Background color.

title
Window title.

antialias
Antialiasing. Can be turned off for somewhat faster drawing.

[Package rkward version 0.7.1 Index]