rk.adjust.system.path {rkward}	R Documentation

Add one or more paths to the filesystem search path used in this session

Description

Add the given path to to the "PATH" environment variable of the running R session. This can be useful to make sure external binaries are found by Sys.which. Paths are normalized before being added, and duplicates are stripped from the path.

Usage

rk.adjust.system.path(add)

Arguments

add
Paths to add. May be missing, in which case the path will no be touched.

Value

A vector of the directories in the file sytem path after the adjustment

[Package rkward version 0.7.1 Index]