rk.call.plugin {rkward}	R Documentation

Call built-in RKWard plugin(s)

Description

rk.call.plugin provides a high level wrapper to call any plugin available in RKWard. The exact string to be used as plugin, and the list of arguments available for a particular plugin, are generally not transparent to the user. rk.list.plugins can be used to obtain a list of current plugins. For plugin arguments, it is recommended to run the plugin, and inspect the "Run again" link that is generated on the output.

Usage

rk.call.plugin(plugin, ..., submit.mode = c("manual", "auto", "submit"))

Arguments

plugin
character string, giving the name of the plugin to call. See Details.

...
arguments passed to the plugin

submit.mode
character string, specifying the submission mode: "manual" will open the plugin GUI and leave it to the user to submit it manually, "auto" will try to submit the plugin, if it can be submitted with the current settings (i.e. if the "Submit"-button is enabled after applying all specified parameters). If the plugin cannot be submitted, with the current settings, it will behave like "manual". "submit" is like "auot", but will close the plugin, and generate an error, if it cannot be submitted. "manual" will always return immediately, "auto" may or may not return immediately, and "submit" will always wait until the plugin has been run, or produce an error.

Details

Warning: Using rk.call.plugin, especially with submit.modes "auto" or "submit" to program a sequence of analyses has important drawbacks. First, the semantics of plugins are not guaranteed to remain unchanged across different versions of RKWard, thus your code may stop working after an upgrade. Second, your code will not be usable outside of an RKWard session. Consider copying the generated code for each plugin, instead. The primary use-cases for rk.call.plugin are automated tests, cross-references, and scripted tutorials.

Note: Even when using "submit.mode=submit", the plugin code is run in the global context. Any local variables of the calling context are not available to the plugin.

Value

rk.call.plugin returns TRUE invisibly.

Author(s)

Thomas Friedrichsmeier rkward-devel@kde.org

See Also

rk.results, rkward://page/rkward_output

Examples

## "t_test_two_vars" plugin:
## see the output: Windows->Show Output
local({
x1 <- rnorm (100)
x2 <- rnorm (100, 2)

rk.call.plugin ("rkward::t_test_two_vars", 
  confint.state="1", conflevel.real="0.95", hypothesis.string="greater", paired.state="0",
      varequal.state="0", 
  x.available="x1", y.available="x2", 
  submit.mode="submit")
})

[Package rkward version 0.7.1 Index]