rk.capture.output {rkward}	R Documentation

Capture output including system commands

Description

This feature records and / or redirects output similar to the sink and capture.output commands provided with R. Contrary to these two, rk.capture.output can also capture output generated by external programs running via system (on both Windows and Unix platforms). Further it allows to capture both warning messages and regular output to the same channel. Output can be passed on for printing, or suppressed, with the default being to print it.

Usage

rk.capture.output(capture.messages = TRUE, capture.output = TRUE,
  suppress.messages = FALSE, suppress.output = FALSE)

rk.end.capture.output(highlight = FALSE)

Arguments

LANG
language code to use. "C" for no translation, i.e. generally English

Details

And arbitrary number of ouptut captures can be started and stopped, and these will form a push/pop stack.

rk.capture.output Starts capturing output.

rk.end.capture.output Stops capturing output, and returns the recorded output in a string.

[Package rkward version 0.7.1 Index]