rk.check.for.pandoc {rkward}	R Documentation

Check for pandoc installation and its features

Description

This function looks for a pandoc installation and if found, gathers information on its path, supported output formats, and the version number.

Usage

rk.check.for.pandoc(stop_if_missing = FALSE, output_format = NA)

Arguments

stop_if_missing
Logical, whether an error should be thrown if either pandoc can't be found at all or a requested output format is not supported.

output_format
A character string, name of a requested output format. If not NA, the function will check if the format is supported and return a boolean value.

Value

A named list with at least the boolean element available, and if that is TRUE, also the character vectors path, output_formats, and version.

Author(s)

Meik Michalke rkward-devel@kde.org

See Also

rk.sessionInfo

Examples

rk.check.for.pandoc()

[Package rkward version 0.7.1 Index]