rk.demo {rkward}	R Documentation

Opens an R demo script for editing

Description

rk.demo behaves similar to demo, but opens the demo script for editing, instead of sourcing it. Contrary to demo, the specification of a topic is mandatory.

Usage

rk.demo(topic, package = NULL, lib.loc = NULL)

Arguments

topic
topic of the example

package
package(s) to search for the demo. If NULL (the default), all currently loaded packages are searched.

lib.loc
Library locations.

Value

Return NULL, unconditionally.

Author(s)

Thomas Friedrichsmeier rkward-devel@kde.org

See Also

rk.edit.files, rk.show.files, demo

Examples

## Not run
rk.demo("graphics")

[Package rkward version 0.7.1 Index]