rk.edit {rkward}	R Documentation

Edit / show an object / file

Description

rk.edit can be used to edit an object in the RKWard data editor. Currently only data.frames are supported. This is similar to edit.data.frame, but the function returns immediately, and the object is edit asynchronously.

Usage

rk.edit(x)

rk.edit.files(name = NULL, file = "", title = NULL, prompt = TRUE)

rk.show.files(file = file, header = file, title = NULL,
  delete.file = FALSE, prompt = TRUE, delete = delete.file)

rk.show.html(url)

Arguments

x
an object to edit.

file
character vector, filenames to show or edit.

title
character vector, of the same length as file; This can be used to give descriptive titles to each file, which will be displayed to the user.

prompt
logical of length 1. If TRUE (the default) a prompt is dialog is shown along with the files to show / edit.

delete
a logical (not NA), when TRUE the shown file(s) are deleted after closing.

wtitle
character vector, of length 1. This will be used as the window title.

Details

rk.edit.files, rk.show.files, and rk.show.html are equivalent to edit, file.show, and browseURL, respectively, but use RKWard as text/html editor/viewer. Generally it is recommended to use edit, file.edit, file.show, and browseURL, instead. These will call the respective RKWard functions by default, when run inside an RKWard session. (via getOption("editor"), and getOption("browser").

Value

All functions described on this page return NULL, unconditionally.

Author(s)

Thomas Friedrichsmeier rkward-devel@kde.org

See Also

edit, file.edit, file.show, browseURL

Examples

## Not run
x <- data.frame (a=c(1:3), b=c(2:4))
rk.edit(x)

[Package rkward version 0.7.1 Index]