rk.list.plugins {rkward}	R Documentation

List of modify loaded plugins

Description

rk.list.plugins returns the a list of all currently registered plugins (in loaded pluginmaps). rk.set.plugin.status allows to control the status of the given plugin(s). Currently, only visibility can be controlled.

Usage

rk.list.plugins()

rk.set.plugin.status(id, context = "", visible = TRUE)

Arguments

id
vector of ids (character) of the plugins to modify

context
in which the plugin should be shown / hidden. This can either be "", meaning the plugin will be affected in all contexts it occurs in, or a character vector of the same length as id.

visible
logical, controlling whether the plugin should be shown (TRUE) or hidden (FALSE). Hidden plugins are essentially removed from the menu. They may still be accessible embedded into other plugins.

Value

rk.list.plugins returns a data.frame listing plugin ids, context, menu path (tab-separated), and label of the plugin. If a plugin is available in more than one context, it will be listed several times. The exact layout (number and order of columns) of this data.frame might be subject to change. However, the names of the columns in the returned data.frame are expected to remain stable. rk.set.plugin.status returns NULL, invisibly

Note: Each call to rk.set.plugin.status will result in a complete rebuild of the menu (in the current implementation). While this should be hardly noticeable in interactive use, it could be an issue when changing the status of many plugins, programatically. In this case, make sure to do all changes in one call to rk.set.plugin.status, rather than many separate calls.

Author(s)

Thomas Friedrichsmeier rkward-devel@kde.org

See Also

rk.call.plugin for invoking a plugin, programatically

Examples

## list all current plugins
rk.list.plugins ()

## NOT RUN
## hide t.test plugin
rk.set.plugin.status ("rkward::t_test", visible=FALSE)
## END NOT RUN

[Package rkward version 0.7.1 Index]