rk.load.pluginmaps {rkward}	R Documentation

(Re-)load the given pluginmap files into the RKWard GUI

Description

(Re-)load the given pluginmap files into the RKWard GUI

Usage

rk.load.pluginmaps(pluginmap.files = NULL, force.add = TRUE,
  force.reload = TRUE)

Arguments

pluginmap.files
a character vector of file names to add. This may be left empty, if the only desired effect is to reload all active pluginmaps.

force.add
logical. Whether the pluginmap files should also be added, if they had been previously de-selected in the settings menu, and regardless of their priority setting. In scripted usage, this should generally be set to FALSE.

force.reload
logical. By default the active pluginmaps are reloaded, only if any new ones were added. If set to TRUE, pluginmaps are reloaded in any case. In scripted usage, this should generally be set to FALSE. NOTE: Since a reload always means reloading _all_ active pluginmaps, This may be slow, and should be used with care.

Note: It is not necessary to reload the pluginmap, in order to refresh an individual plugin (e.g. after modifying the dialog), as plugins are not kept in memory after closing. Any currently opened plugins are not affected by this function.

Author(s)

Thomas Friedrichsmeier rkward-devel@kde.org

See Also

rk.call.plugin, @seealso rkwarddev::rk.plugin.skeleton

Examples

## NOT RUN

## reload all active pluginmaps
rk.load.pluginmaps()

## END NOT RUN

[Package rkward version 0.7.1 Index]