rk.rename.in.container {rkward}	R Documentation

Miscellaneous utility functions

Description

rk.rename.in.container renames a named object (column/element) in a data.frame/list without changing its position.

Usage

rk.rename.in.container(x, old_name, new_name, envir = parent.frame())

rk.make.repos.string()

rk.select.CRAN.mirror()

Arguments

x
a data.frame or list.

old_name
a string, the name of the column or element to be renamed.

new_name
a string, the new name.

envir
an environment where x is available.

Details

rk.make.repos.string just creates a R statement for repos. A typical user should not need to use this function.

rk.select.CRAN.mirror is an in-house replacement for chooseCRANmirror without changing options ("repos"), permanently. It uses native KDE gui and provides more information on each mirror.

Value

rk.rename.in.container returns NULL on successfule renaming, otherwise an error.

rk.make.repos.string returns a valid R expression as a character string which can then be parsed and evaluated.

rk.select.CRAN.mirror returns the URL of the chosen mirror, as a string.

Author(s)

Thomas Friedrichsmeier rkward-devel@kde.org

Examples

## rk.rename.in.container
ir <- iris
str (ir)
rk.rename.in.container(ir, "Species", "Taxonomic.Group")
str (ir)

[Package rkward version 0.7.1 Index]