rk.printer.device {rkward}	R Documentation

Device for printing using the KDE print dialog

Description

Creates a device operating on a temporary file (internally a postscript() device). When the device is closed, it is printed, automatically, using the KDE print dialog (if installed).

Usage

rk.printer.device(...)

Arguments

...
arguments are passed to postscript

Details

Typically this device is used with dev.print, as shown in the example, below.

Value

Returns the name of the underlying temporary file, invisibly.

Author(s)

Thomas Friedrichsmeier rkward-devel@kde.org

See Also

postscript, dev.print, rk.graph.on

Examples

## Not run:
plot (rnorm (10))
dev.print (rk.printer.device)

[Package rkward version 0.7.1 Index]