rk.record.commands {rkward}	R Documentation

Start recording commands that are submitted from RKWard to R

Description

To stop recording, supply NULL or "" as filename. Currently used for the purpose of automated testing, only. Perhaps in the future this or a similar mechanism could also be added as a user feature.

Usage

rk.record.commands(filename, include.all = FALSE)

Arguments

filename
filename to write to (file will be truncated!).

include.all
By default, some types of command are filtered (internal synchronisation commands, and run again links). Should these be included?

[Package rkward version 0.7.1 Index]