rk.sessionInfo {rkward}	R Documentation

Print information on the RKWard session

Description

Gathers and prints information on the setup of the current RKWard session. In general, you should always include this information when reporting a bug in RKWard.

Usage

rk.sessionInfo()

Details

Typically, when reporting a bug, you should use Help->Report Bug... from the menu. Internally, this will call rk.sessionInfo().

Value

Returns the object created by sessionInfo(), invisibly. Note that this includes only the information on the R portion of the session.

Author(s)

Thomas Friedrichsmeier rkward-devel@kde.org

See Also

sessionInfo

Examples

rk.sessionInfo()

[Package rkward version 0.7.1 Index]