rk.show.message {rkward}	R Documentation

Message boxes and selection list using native KDE GUI

Description

Multi-purpose pop-up message boxes and selection list using native KDE GUI elements. The message boxes can be used either to show some information or ask some question. The selection list can be used to get a vector of selected items.

Usage

rk.show.message(message, caption = gettext("Information"), wait = TRUE)

rk.show.question(message, caption = gettext("Question"),
  button.yes = "yes", button.no = "no", button.cancel = "cancel",
  default = TRUE)

rk.askYesNo(msg, default = TRUE, prompts = c("yes", "no", "cancel"),
  caption = gettext("Question"), is.rk.askYesNo = TRUE, ...)

rk.select.list(list, preselect = NULL, multiple = FALSE,
  title = NULL)

Arguments

message
a string for the content of the message box.

caption
a string for title of the message box.

wait
a logical (not NA) indicating whether the R interpreter should wait for the user's action, or run it asynchronously.

button.yes
a string for the text label of the Yes button. Can be an empty string (""), in which case the button is not displayed at all. Note that the default value of "yes" (lower case) means to use a default rendering of a "Yes" button, which may or may not be the literal string "Yes" (importantly, it will be translated, automatically).

button.no
a string used for the text label of the No button. This behaves similar to button.yes, including the meaning of the default value "no".

button.cancel
a string used for the text label of the Cancel button. This behaves similar to button.yes, including the meaning of the default value "cancel".

default
The expected or "safe" default response (e.g. TRUE for "Yes button"). The corresponding button will focused, so that it will become selected option, if the user simply presses enter.

msg
like message, only the argument was renamed to mimic the formals of askYesNo.

prompts
either a character vector containing 3 prompts corresponding to return values of TRUE, FALSE, or NA, or a single character value containing the prompts separated by / characters.

is.rk.askYesNo
a logical value, you can safely ignore this argument if you call rk.askYesNo manually. This argument is needed if rk.askYesNo is set via options("askYesNo"=rk.askYesNo) because otherwise we'd either need more complicated function code there, fail with an error or end up in an infinite loop.

list
a vector, coerced into a character vector.

multiple
a logical (not NA), when TRUE multiple selection selection is allowed.

title
a string, for the window title of the displayed list

preselct
a vector, coerced into a character vector, items to be preselected.

Details

For rk.show.question, the R interpreter always waits for the user's choice.

rk.select.list replaces utils::select.list for the running session acting as a drop-in replacement for tk_select.list. Use .rk.backups$select.list for the original utils::select.list function (see Examples).

Value

rk.show.message always returns TRUE, invisibly.

rk.show.question returns TRUE for Yes, FALSE for No, and NULL for Cancel actions. If the dialog is closed without clicking on any button, NULL is returned, as well.

rk.askYesNo has the same return values as rk.show.question, except it returns NA for Cancel actions.

rk.select.list returns the value of select.list.

Author(s)

Thomas Friedrichsmeier rkward-devel@kde.org

See Also

system, select.list

Examples

require (rkward)

## Message boxes
if (rk.show.question ("Question:\nDo you want to know about RKWard?", 
    button.yes = "Yes, I do!", button.no = "No, I don't care!", button.cancel = "")) {
  rk.show.message ("Message:\nRKWard is a KDE GUI for R.", "RKWard Info")
} else {
  rk.show.message ("You must be joking!", "RKWard Info",
      wait = FALSE) ## Run asynchronously
}

## Selection lists:
rk.select.list (LETTERS, preselect = c("A", "E", "I", "O", "U"), 
  multiple = TRUE, title = "vowels")
.rk.backups$select.list (LETTERS, preselect = c("A", "E", "I", "O", "U"), 
  multiple = TRUE, title = "vowels")

[Package rkward version 0.7.1 Index]