rk.switch.frontend.language {rkward}	R Documentation

Switch language / translation to use in the frontend.

Description

This feature is mostly intended for the purpose of automated testing, which needs a defined language to work. It might also be useful for translators, or e.g. to look up some terms untranslated in special cases where you are more familiar with the English terms than your native language terms. Note that this will only strings that are translated after the call., only those which get translated after the call. Most new dialogs you open, and importantly new plugin dialogs should show strings in the new lanuage, however.

Usage

rk.switch.frontend.language(LANG = "C")

Arguments

LANG
language code to use. "C" for no translation, i.e. generally English

Details

To change the language in the backend, use Sys.setenv(LANGUAGE=...) or Sys.setlocale().

[Package rkward version 0.7.1 Index]