rk.sync {rkward}	R Documentation

Sync R object(s)

Description

RKWard keeps an internal representation of objects in the R workspace. For objects in the .GlobalEnv, this representation is updated after each top-level statement. For the rare cases where this is not enough, rk.sync can be used to update the representation of a single object, x, while rk.sync.global scans the .GlobalEnv for new and removed objects, and updates as appropriate.

Usage

rk.sync(x)

rk.sync.global()

Arguments

x
any R object to sync

Details

These functions are rarely needed outside automated testing. However, rk.sync() can be useful, if an object outside the .GlobalEnv has changed, since this will not be detected automatically. Also, by default RKWard does not recurse into environments when updating its representation of objects. rk.sync() can be used, here, to inspect the objects inside environments (see examples).

Value

NULL, invisibly.

Author(s)

Thomas Friedrichsmeier rkward-devel@kde.org

See Also

rkward://page/rkward_workspace_browser

Examples

rk.sync (rkward::rk.record.plot)

[Package rkward version 0.7.1 Index]