rk.without.plot.history {rkward}	R Documentation

Run a (plotting) action, without recording anything in the plot history. Internally, the plot history option is turned off for the duration of the action.

Description

Run a (plotting) action, without recording anything in the plot history. Internally, the plot history option is turned off for the duration of the action.

Usage

rk.without.plot.history(expr)

[Package rkward version 0.7.1 Index]