rk.save.workplace {rkward}	R Documentation

Save or restore RKWard workplace

Description

rk.save.workplace can be used to save a representation of the RKWard workplace (i.e. which scripts, data edtiors and other windows are shown) to a file. rk.restore.workplace restores an RKWard workplace as saved by rk.save.workplace.

Usage

rk.save.workplace(file = NULL, description = NULL)

rk.restore.workplace(file = NULL, close.windows = TRUE)

Arguments

file
a character string giving the url of the file to save to, or NULL for automatic selection of a suitable file (see Details).

description
For internal use, only. A character string describing the workplace status to save. Generally, you should leave this as the default value (NULL).

close.windows
a logical; whether current windows should be closed before restoring.

Details

If the file parameter is omitted (or NULL), a suitable filename is selected automatically. If a workspace has been loaded, this is the URL of the workspace with an appended .rkworkplace. Otherwise a filename in the RKWard directory, as generated by rk.get.tempfile.name.

NOTE: Not all types of windows can be saved and restored. Esp. graphics device windows will not be restored (but WILL be closed by rk.restore.workplace(), if close.windows is TRUE).

Value

Both functions return NULL.

Author(s)

Thomas Friedrichsmeier rkward-devel@kde.org

See Also

rkward://page/rkward_for_r_users, rk.get.workspace.url

Examples

## Not run
rk.save.workplace ()
rk.restore.workplace ()
## End not run

[Package rkward version 0.7.1 Index]