rkward-package {rkward}	R Documentation

Provides functions related to the RKWard GUI

Description

This package contains functions which are useful in combination with the RKWard GUI. Many of these functions only needed for the internal communication between RKWard and R, but some are also useful in user scripts.

Details

The DESCRIPTION file:

Package: 	 rkward

Type: 	 Package

Version: 	 0.7.1

Date: 	 2018-10-11

Depends: 	 R (>= 2.9.0),methods

Encoding: 	 UTF-8

License: 	 GPL (>= 2)

LazyLoad: 	 yes

URL: 	 https://rkward.kde.org

Author(s)

Thomas Friedrichsmeier and the RKWard Team

Maintainer: RKWard-devel mailing list <rkward-devel@kde.org>

See Also

Useful links:

https://rkward.kde.org

Report bugs at https://rkward.kde.org/Bugs

[Package rkward version 0.7.1 Index]